plot tabs skeleton, props
