+++ src/components/Differential/SVGPlot.scss
@@ -6,14 +6,6 @@
   display: flex;
   flex-direction: column;
 
-  .svgPlotDropdown {
-    max-width: fit-content !important;
-    z-index: 5 !important;
-    margin-top: 0px;
-    margin-bottom: 10px !important;
-    margin-left: 40px !important;
-  }
-
   .ui.secondary.pointing.menu {
     margin-left: 35px !important;
   }
@@ -69,21 +61,27 @@
   cursor: pointer !important;
 }
 
+.svgPlotDropdown {
+  max-width: fit-content !important;
+  z-index: 5 !important;
+  margin-top: 0px;
+  margin-bottom: 10px !important;
+  margin-left: 40px !important;
+}
+
 .svgPlotDropdownInOverlay {
   max-width: fit-content !important;
   z-index: 5 !important;
   margin: 10px 10px 20px 10px !important;
 }
 
-#svgPlotDropdownSingle {
-  left: -35px !important;
-}
 #svgPlotDropdownDifferential {
   left: -50px !important;
   top: 8px;
   // border-bottom: 2px solid var(--color-link);
 }
 
+#svgPlotDropdownSingle, 
 #svgPlotDropdownMulti {
   left: -35px !important;
   top: 8px;
